Set random run chance
To accomodate course scheduling options with Heroku scheduler, gave a
configurable chance for the program to actually run. That way it can run
multiple times a day without overposting.

+++ bot.py
@@ -1,18 +1,20 @@
 import tweepy
 import os
+from random import randint
 from image_generator import generateImage
 
-API_KEY             = os.environ['API_KEY']
-API_SECRET          = os.environ['API_SECRET']
-ACCESS_TOKEN        = os.environ['ACCESS_TOKEN']
-ACCESS_TOKEN_SECRET = os.environ['ACCESS_TOKEN_SECRET']
+if randint(1, int(os.environ['RUN_CHANCE'])) == 1:
+    API_KEY             = os.environ['API_KEY']
+    API_SECRET          = os.environ['API_SECRET']
+    ACCESS_TOKEN        = os.environ['ACCESS_TOKEN']
+    ACCESS_TOKEN_SECRET = os.environ['ACCESS_TOKEN_SECRET']
 
-auth = tweepy.OAuthHandler(API_KEY, API_SECRET)
-auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)
-api = tweepy.API(auth)
+    auth = tweepy.OAuthHandler(API_KEY, API_SECRET)
+    auth.set_access_token(ACCESS_TOKEN, ACCESS_TOKEN_SECRET)
+    api = tweepy.API(auth)
 
-imagePath, (quranFooter, bibleFooter) = generateImage()
-statusText = '{0}\n{1}'.format(quranFooter, bibleFooter)
+    imagePath, (quranFooter, bibleFooter) = generateImage()
+    statusText = '{0}\n{1}'.format(quranFooter, bibleFooter)
 
-api.update_with_media(imagePath, statusText)
-os.remove(imagePath)
+    api.update_with_media(imagePath, statusText)
+    os.remove(imagePath)
